+++ ta4j-core/src/main/java/org/ta4j/core/indicators/DPOIndicator.java
@@ -33,6 +33,7 @@ import org.ta4j.core.num.Num;
 /**
  * The Detrended Price Oscillator (DPO) indicator.
  *
+ * <p>
  * The Detrended Price Oscillator (DPO) is an indicator designed to remove trend
  * from price and make it easier to identify cycles. DPO does not extend to the
  * last date because it is based on a displaced moving average. However,
@@ -40,6 +41,7 @@ import org.ta4j.core.num.Num;
  * oscillator. Instead, DPO is used to identify cycles highs/lows and estimate
  * cycle length.
  *
+ * <p>
  * In short, DPO(20) equals price 11 days ago less the 20-day SMA.
  *
  * @see <a href=
@@ -54,7 +56,7 @@ public class DPOIndicator extends CachedIndicator<Num> {
     /**
      * Constructor.
      *
-     * @param series   the series
+     * @param series   the bar series
      * @param barCount the time frame
      */
     public DPOIndicator(BarSeries series, int barCount) {