+++ ta4j-core/src/main/java/org/ta4j/core/indicators/aroon/AroonFacade.java
@@ -35,23 +35,24 @@ import org.ta4j.core.indicators.numeric.NumericIndicator;
  * not cached, although they may be wrapped around cached objects.
  */
 public class AroonFacade {
+
     private final NumericIndicator up;
     private final NumericIndicator down;
 
     /**
      * Create the Aroon facade.
-     * 
-     * @param bs a bar series
-     * @param n  the number of periods (barCount) used for the indicators
+     *
+     * @param series   the bar series
+     * @param barCount the number of periods used for the indicators
      */
-    public AroonFacade(BarSeries bs, int n) {
-        this.up = NumericIndicator.of(new AroonUpIndicator(bs, n));
-        this.down = NumericIndicator.of(new AroonDownIndicator(bs, n));
+    public AroonFacade(BarSeries series, int barCount) {
+        this.up = NumericIndicator.of(new AroonUpIndicator(series, barCount));
+        this.down = NumericIndicator.of(new AroonDownIndicator(series, barCount));
     }
 
     /**
      * A fluent AroonUp indicator.
-     * 
+     *
      * @return a NumericIndicator wrapped around a cached AroonUpIndicator
      */
     public NumericIndicator up() {
@@ -60,7 +61,7 @@ public class AroonFacade {
 
     /**
      * A fluent AroonDown indicator.
-     * 
+     *
      * @return a NumericIndicator wrapped around a cached AroonDownIndicator
      */
     public NumericIndicator down() {
@@ -69,7 +70,7 @@ public class AroonFacade {
 
     /**
      * A lightweight fluent AroonOscillator.
-     * 
+     *
      * @return an uncached object that calculates the difference between AoonUp and
      *         AroonDown
      */