+++ ta4j-core/src/main/java/org/ta4j/core/indicators/bollinger/BollingerBandsMiddleIndicator.java
@@ -29,9 +29,11 @@ import org.ta4j.core.num.Num;
 
 /**
  * Buy - Occurs when the price line crosses from below to above the Lower
- * Bollinger Band. Sell - Occurs when the price line crosses from above to below
- * the Upper Bollinger Band.
+ * Bollinger Band.
  * 
+ * <p>
+ * Sell - Occurs when the price line crosses from above to below the Upper
+ * Bollinger Band.
  */
 public class BollingerBandsMiddleIndicator extends CachedIndicator<Num> {
 
@@ -39,9 +41,8 @@ public class BollingerBandsMiddleIndicator extends CachedIndicator<Num> {
 
     /**
      * Constructor.
-     * 
-     * @param indicator the indicator that gives the values of the middle band
      *
+     * @param indicator the indicator that gives the values of the middle band
      */
     public BollingerBandsMiddleIndicator(Indicator<Num> indicator) {
         super(indicator);