+++ ta4j-core/src/main/java/org/ta4j/core/indicators/donchian/DonchianChannelMiddleIndicator.java
@@ -32,13 +32,18 @@ import org.ta4j.core.num.Num;
  */
 public class DonchianChannelMiddleIndicator extends CachedIndicator<Num> {
 
+    private final int barCount;
     private final DonchianChannelLowerIndicator lower;
     private final DonchianChannelUpperIndicator upper;
-    private final int barCount;
 
+    /**
+     * Constructor.
+     * 
+     * @param series   the bar series
+     * @param barCount the time frame
+     */
     public DonchianChannelMiddleIndicator(BarSeries series, int barCount) {
         super(series);
-
         this.barCount = barCount;
         this.lower = new DonchianChannelLowerIndicator(series, barCount);
         this.upper = new DonchianChannelUpperIndicator(series, barCount);