+++ ta4j-core/src/main/java/org/ta4j/core/indicators/helpers/HighestValueIndicator.java
@@ -29,13 +29,21 @@ import org.ta4j.core.num.Num;
 
 /**
  * Highest value indicator.
+ * 
+ * <p>
+ * Returns the highest indicator value from the bar series within the bar count.
  */
 public class HighestValueIndicator extends CachedIndicator<Num> {
 
     private final Indicator<Num> indicator;
-
     private final int barCount;
 
+    /**
+     * Constructor.
+     * 
+     * @param indicator the {@link Indicator}
+     * @param barCount  the time frame
+     */
     public HighestValueIndicator(Indicator<Num> indicator, int barCount) {
         super(indicator);
         this.indicator = indicator;